feat: 키보드 설정으로 인해 발생한 cycle 오류 해결 #3
